Separate remote tests
